+++ tests/WidgetTest/DbCacheTest.php
@@ -18,4 +18,27 @@ class DbCacheTest extends CacheTestCase
         @unlink('cache.sqlite');
         parent::tearDownAfterClass();
     }
+
+    public function testExpire()
+    {
+        $cache = $this->object;
+        $key = __FUNCTION__;
+
+        $result = $cache->set($key, true, -1);
+        $this->assertTrue($result);
+
+        //sleep(2);
+
+        $result = $cache->exists($key);
+        $this->assertFalse($result);
+
+        $result = $cache->set($key, true, -1);
+        $this->assertTrue($result);
+
+        //sleep(2);
+
+        $result = $cache->get($key);
+        $this->assertFalse($result);
+    }
+
 }